requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
Django==6.0.4
djangorestframework==3.17.1
pillow==12.2.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
